+++ tests/e2e/global-teardown.ts
@@ -1,0 +1,7 @@
+import { Database } from './helpers/database';
+
+async function globalTeardown() {
+  await Database.cleanup();
+}
+
+export default globalTeardown;